# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/orders": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_getOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/ready": {
      "patch": {
        "operationId": "OrdersController_markOrderReady",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}/closed": {
      "patch": {
        "operationId": "OrdersController_markOrderClosed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/floors": {
      "get": {
        "operationId": "FloorController_getFloors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список этажей",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FloorDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получение списка этажей с офисами",
        "tags": [
          "Floors"
        ]
      },
      "post": {
        "operationId": "FloorController_createFloor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFloorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Этаж успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создание нового этажа",
        "tags": [
          "Floors"
        ]
      }
    },
    "/floors/{id}": {
      "put": {
        "operationId": "FloorController_updateFloor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFloorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Этаж успешно обновлён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Редактирование этажа",
        "tags": [
          "Floors"
        ]
      },
      "delete": {
        "operationId": "FloorController_deleteFloor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Этаж успешно удалён"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удаление этажа",
        "tags": [
          "Floors"
        ]
      }
    },
    "/floors/office": {
      "post": {
        "operationId": "FloorController_createOffice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOfficeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Офис успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создание офиса на этаже",
        "tags": [
          "Floors"
        ]
      }
    },
    "/floors/office/{id}": {
      "put": {
        "operationId": "FloorController_updateOffice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOfficeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Офис успешно обновлён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Редактирование офиса",
        "tags": [
          "Floors"
        ]
      },
      "delete": {
        "operationId": "FloorController_deleteOffice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Офис успешно удалён"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удаление офиса",
        "tags": [
          "Floors"
        ]
      }
    }
  },
  "info": {
    "title": "Shawarma Order Notification System",
    "description": "API documentation for Shawarma Order Notification System",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "FloorDto": {
        "type": "object",
        "properties": {}
      },
      "CreateFloorDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateFloorDto": {
        "type": "object",
        "properties": {}
      },
      "CreateOfficeDto": {
        "type": "object",
        "properties": {}
      },
      "OfficeDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateOfficeDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}